chore: move v2-to-v3 add utils to utils/add
