+++ src/transforms/v2-to-v3/utils/add/addV3ClientImport.ts
@@ -1,8 +1,8 @@
 import { Collection, JSCodeshift } from "jscodeshift";
 
+import { PACKAGE_NAME } from "../config";
+import { getV2ServiceModulePath } from "../get";
 import { AddV3ClientModuleOptions } from "./addV3ClientModule";
-import { PACKAGE_NAME } from "./config";
-import { getV2ServiceModulePath } from "./get";
 
 export const addV3ClientImport = (
   j: JSCodeshift,
