+++ src/transforms/v2-to-v3/utils/add/addV3ClientRequire.ts
@@ -1,8 +1,8 @@
 import { Collection, Identifier, JSCodeshift, ObjectPattern, Property } from "jscodeshift";
 
+import { PACKAGE_NAME } from "../config";
+import { getRequireVariableDeclaration, getV2ServiceModulePath } from "../get";
 import { AddV3ClientModuleOptions } from "./addV3ClientModule";
-import { PACKAGE_NAME } from "./config";
-import { getRequireVariableDeclaration, getV2ServiceModulePath } from "./get";
 
 export const addV3ClientRequire = (
   j: JSCodeshift,
